шаблонизация и фильтрация
