+++ config.php
@@ -1,0 +1,6 @@
+<?php
+
+$config = [
+	// сайт доступен и работает
+	'enable' => true
+];
